fix: rename file for true-false widget
