+++ src/components/widgets/true-false/true-false-strategy.scss
